requirements.txt
# Scraping
requests
playwright
markdownify

# YouTube
youtube-transcript-api
google-api-python-client

# Utilities
python-dotenv     # for managing credentials locally
python-frontmatter   # for adding frontmatter to markdown files